refactor: moo-ve sellCow to its own file
